+++ R/wflow_build.R
@@ -213,6 +213,8 @@ wflow_build <- function(files = NULL, make = is.null(files),
   if (!(is.character(project) && length(project) == 1))
     stop("project must be a one-element character vector")
 
+  check_wd_exists()
+
   if (!fs::dir_exists(project)) {
     stop("project directory does not exist.")
   }