+++ R/wflow_git_commit.R
@@ -112,6 +112,8 @@ wflow_git_commit <- function(files = NULL, message = NULL, all = FALSE,
   if (!(is.character(project) && length(project) == 1))
     stop("project must be a one-element character vector")
 
+  check_wd_exists()
+
   if (!fs::dir_exists(project)) {
     stop("project directory does not exist.")
   }