+++ R/wflow_git_push.R
@@ -121,6 +121,8 @@ wflow_git_push <- function(remote = NULL, branch = NULL, username = NULL,
   if (!(is.character(project) && length(project) == 1))
     stop("project must be a one-element character vector")
 
+  check_wd_exists()
+
   if (!fs::dir_exists(project)) {
     stop("project directory does not exist.")
   }