+++ R/wflow_rename.R
@@ -93,6 +93,8 @@ wflow_rename <- function(files,
   if (!(is.character(project) && length(project) == 1))
     stop("project must be a one-element character vector")
 
+  check_wd_exists()
+
   if (!fs::dir_exists(project)) {
     stop("project directory does not exist.")
   }