+++ R/wflow_rename_proj.R
@@ -80,6 +80,8 @@ wflow_rename_proj <- function(name,
   if (!(is.character(project) && length(project) == 1))
     stop("project must be a one-element character vector")
 
+  check_wd_exists()
+
   if (!fs::dir_exists(project)) {
     stop("project directory does not exist.")
   }