+++ R/wflow_use_github.R
@@ -154,6 +154,8 @@ wflow_use_github <- function(username = NULL, repository = NULL,
   if (!(is.character(project) && length(project) == 1))
     stop("project must be a one-element character vector")
 
+  check_wd_exists()
+
   if (!fs::dir_exists(project)) {
     stop("project directory does not exist.")
   }