+++ R/wflow_use_gitlab.R
@@ -100,6 +100,8 @@ wflow_use_gitlab <- function(username = NULL, repository = NULL,
   if (!(is.character(project) && length(project) == 1))
     stop("project must be a one-element character vector")
 
+  check_wd_exists()
+
   if (!fs::dir_exists(project)) {
     stop("project directory does not exist.")
   }